Revert "Taz branch2"
