--- Read and Tell.xcworkspace/xcuserdata/josephsiggia.xcuserdatad/xcdebugger/Breakpoints_v2.xcbkptlist
@@ -1,116 +1,0 @@
-<?xml version="1.0" encoding="UTF-8"?>
-<Bucket
-   uuid = "FD08DB6E-27AB-4586-B57C-E0C02B047232"
-   type = "0"
-   version = "2.0">
-   <Breakpoints>
-      <BreakpointProxy
-         BreakpointExtensionID = "Xcode.Breakpoint.FileBreakpoint">
-         <BreakpointContent
-            uuid = "A5440482-A696-4ACC-AA34-9BC4764EA6B4"
-            shouldBeEnabled = "No"
-            ignoreCount = "0"
-            continueAfterRunningActions = "No"
-            filePath = "Read and Tell/ReviewsContentViewController.swift"
-            startingColumnNumber = "9223372036854775807"
-            endingColumnNumber = "9223372036854775807"
-            startingLineNumber = "17"
-            endingLineNumber = "17"
-            landmarkName = "ReviewsContentViewController"
-            landmarkType = "3">
-            <Locations>
-               <Location
-                  uuid = "A5440482-A696-4ACC-AA34-9BC4764EA6B4 - 2856958611348fbe"
-                  shouldBeEnabled = "Yes"
-                  ignoreCount = "0"
-                  continueAfterRunningActions = "No"
-                  symbolName = "Read_and_Tell.ReviewsContentViewController.reviewTableView.getter : Swift.Optional&lt;__C.UITableView&gt;"
-                  moduleName = "Read and Tell"
-                  usesParentBreakpointCondition = "Yes"
-                  urlString = "file:///Users/josephsiggia/Desktop/codepath-group-project-BookApp/Read%20and%20Tell/ReviewsContentViewController.swift"
-                  startingColumnNumber = "9223372036854775807"
-                  endingColumnNumber = "9223372036854775807"
-                  startingLineNumber = "17"
-                  endingLineNumber = "17"
-                  offsetFromSymbolStart = "88">
-               </Location>
-               <Location
-                  uuid = "A5440482-A696-4ACC-AA34-9BC4764EA6B4 - 4e1b35c1222d3252"
-                  shouldBeEnabled = "Yes"
-                  ignoreCount = "0"
-                  continueAfterRunningActions = "No"
-                  symbolName = "Read_and_Tell.ReviewsContentViewController.reviewTableView.setter : Swift.Optional&lt;__C.UITableView&gt;"
-                  moduleName = "Read and Tell"
-                  usesParentBreakpointCondition = "Yes"
-                  urlString = "file:///Users/josephsiggia/Desktop/codepath-group-project-BookApp/Read%20and%20Tell/ReviewsContentViewController.swift"
-                  startingColumnNumber = "9223372036854775807"
-                  endingColumnNumber = "9223372036854775807"
-                  startingLineNumber = "17"
-                  endingLineNumber = "17"
-                  offsetFromSymbolStart = "108">
-               </Location>
-               <Location
-                  uuid = "A5440482-A696-4ACC-AA34-9BC4764EA6B4 - 9b8921b84c0726db"
-                  shouldBeEnabled = "Yes"
-                  ignoreCount = "0"
-                  continueAfterRunningActions = "No"
-                  symbolName = "Read_and_Tell.ReviewsContentViewController.reviewTableView.modify : Swift.Optional&lt;__C.UITableView&gt;"
-                  moduleName = "Read and Tell"
-                  usesParentBreakpointCondition = "Yes"
-                  urlString = "file:///Users/josephsiggia/Desktop/codepath-group-project-BookApp/Read%20and%20Tell/ReviewsContentViewController.swift"
-                  startingColumnNumber = "9223372036854775807"
-                  endingColumnNumber = "9223372036854775807"
-                  startingLineNumber = "17"
-                  endingLineNumber = "17"
-                  offsetFromSymbolStart = "0">
-               </Location>
-               <Location
-                  uuid = "A5440482-A696-4ACC-AA34-9BC4764EA6B4 - cb7218168d5a0bfe"
-                  shouldBeEnabled = "Yes"
-                  ignoreCount = "0"
-                  continueAfterRunningActions = "No"
-                  symbolName = "Read_and_Tell.ReviewsContentViewController.reviewTableView.modify : Swift.Optional&lt;__C.UITableView&gt; with unmangled suffix &quot;.resume.0&quot;"
-                  moduleName = "Read and Tell"
-                  usesParentBreakpointCondition = "Yes"
-                  urlString = "file:///Users/josephsiggia/Desktop/codepath-group-project-BookApp/Read%20and%20Tell/ReviewsContentViewController.swift"
-                  startingColumnNumber = "9223372036854775807"
-                  endingColumnNumber = "9223372036854775807"
-                  startingLineNumber = "17"
-                  endingLineNumber = "17"
-                  offsetFromSymbolStart = "28">
-               </Location>
-               <Location
-                  uuid = "A5440482-A696-4ACC-AA34-9BC4764EA6B4 - 73e145b2652b313"
-                  shouldBeEnabled = "Yes"
-                  ignoreCount = "0"
-                  continueAfterRunningActions = "No"
-                  symbolName = "Read_and_Tell.ReviewsContentViewController.init(nibName: Swift.Optional&lt;Swift.String&gt;, bundle: Swift.Optional&lt;__C.NSBundle&gt;) -&gt; Read_and_Tell.ReviewsContentViewController"
-                  moduleName = "Read and Tell"
-                  usesParentBreakpointCondition = "Yes"
-                  urlString = "file:///Users/josephsiggia/Desktop/codepath-group-project-BookApp/Read%20and%20Tell/ReviewsContentViewController.swift"
-                  startingColumnNumber = "9223372036854775807"
-                  endingColumnNumber = "9223372036854775807"
-                  startingLineNumber = "17"
-                  endingLineNumber = "17"
-                  offsetFromSymbolStart = "96">
-               </Location>
-               <Location
-                  uuid = "A5440482-A696-4ACC-AA34-9BC4764EA6B4 - bf30c02f3aa3b2ef"
-                  shouldBeEnabled = "Yes"
-                  ignoreCount = "0"
-                  continueAfterRunningActions = "No"
-                  symbolName = "Read_and_Tell.ReviewsContentViewController.init(coder: __C.NSCoder) -&gt; Swift.Optional&lt;Read_and_Tell.ReviewsContentViewController&gt;"
-                  moduleName = "Read and Tell"
-                  usesParentBreakpointCondition = "Yes"
-                  urlString = "file:///Users/josephsiggia/Desktop/codepath-group-project-BookApp/Read%20and%20Tell/ReviewsContentViewController.swift"
-                  startingColumnNumber = "9223372036854775807"
-                  endingColumnNumber = "9223372036854775807"
-                  startingLineNumber = "17"
-                  endingLineNumber = "17"
-                  offsetFromSymbolStart = "72">
-               </Location>
-            </Locations>
-         </BreakpointContent>
-      </BreakpointProxy>
-   </Breakpoints>
-</Bucket>
