+++ Read and Tell/Base.lproj/Main.storyboard
@@ -439,14 +439,14 @@
         <scene sceneID="BUS-cq-rRl">
             <objects>
                 <viewController id="cY5-vL-hmJ" customClass="SearchBookDetailViewController" customModule="Read_and_Tell" customModuleProvider="target" sceneMemberID="viewController">
-                    <view key="view" contentMode="scaleToFill" ambiguous="YES" id="u61-ah-ktO">
+                    <view key="view" contentMode="scaleToFill" id="u61-ah-ktO">
                         <rect key="frame" x="0.0" y="0.0" width="414" height="896"/>
                         <autoresizingMask key="autoresizingMask" widthSizable="YES" heightSizable="YES"/>
                         <subviews>
-                            <scrollView clipsSubviews="YES" multipleTouchEnabled="YES" contentMode="scaleToFill" ambiguous="YES" translatesAutoresizingMaskIntoConstraints="NO" id="Ue4-tr-wXW">
+                            <scrollView clipsSubviews="YES" multipleTouchEnabled="YES" contentMode="scaleToFill" translatesAutoresizingMaskIntoConstraints="NO" id="Ue4-tr-wXW">
                                 <rect key="frame" x="0.0" y="0.0" width="414" height="896"/>
                                 <subviews>
-                                    <view contentMode="scaleToFill" ambiguous="YES" translatesAutoresizingMaskIntoConstraints="NO" id="lfg-ge-22y">
+                                    <view contentMode="scaleToFill" translatesAutoresizingMaskIntoConstraints="NO" id="lfg-ge-22y">
                                         <rect key="frame" x="0.0" y="0.0" width="414" height="1500"/>
                                         <subviews>
                                             <imageView clipsSubviews="YES" userInteractionEnabled="NO" contentMode="scaleAspectFit" horizontalHuggingPriority="251" verticalHuggingPriority="251" fixedFrame="YES" translatesAutoresizingMaskIntoConstraints="NO" id="vOD-sA-2f7">
@@ -546,7 +546,8 @@
                                                 <state key="normal" title="Button"/>
                                                 <buttonConfiguration key="configuration" style="plain"/>
                                                 <connections>
-                                                    <action selector="reviewButton:" destination="cY5-vL-hmJ" eventType="touchUpInside" id="tVl-8A-sSA"/>
+                                                    <action selector="reviewsButton:" destination="cY5-vL-hmJ" eventType="touchUpInside" id="qdm-ai-hDz"/>
+                                                    <segue destination="wwr-ya-60d" kind="show" id="h1Z-ta-4aP"/>
                                                 </connections>
                                             </button>
                                         </subviews>
@@ -576,11 +577,7 @@
                         </constraints>
                     </view>
                     <navigationItem key="navigationItem" id="Gqn-sh-Tla">
-                        <barButtonItem key="rightBarButtonItem" title="Save to Collection" id="dnM-Je-2Mr">
-                            <connections>
-                                <action selector="saveToCollection:" destination="cY5-vL-hmJ" id="HJW-xz-KJb"/>
-                            </connections>
-                        </barButtonItem>
+                        <barButtonItem key="rightBarButtonItem" title="Save to Collection" id="dnM-Je-2Mr"/>
                     </navigationItem>
                     <size key="freeformSize" width="414" height="896"/>
                     <connections>
@@ -594,19 +591,18 @@
                         <outlet property="star4" destination="a92-8s-nOR" id="UWq-UF-M69"/>
                         <outlet property="star5" destination="tBO-BZ-GmG" id="HlV-dP-kn6"/>
                         <outlet property="synopsisLabel" destination="7DA-q2-Gku" id="I3I-Fb-vIH"/>
-                        <segue destination="wwr-ya-60d" kind="presentation" identifier="toReviewsSegue" id="AKc-Xz-LRj"/>
                     </connections>
                 </viewController>
                 <placeholder placeholderIdentifier="IBFirstResponder" id="A8d-WY-gpv" userLabel="First Responder" customClass="UIResponder" sceneMemberID="firstResponder"/>
             </objects>
             <point key="canvasLocation" x="100.00000000000001" y="1135.0446428571429"/>
         </scene>
-        <!--Reviews Content View Controller-->
+        <!--Reviews View Controller-->
         <scene sceneID="XBj-ER-Ifj">
             <objects>
-                <viewController id="wwr-ya-60d" customClass="ReviewsContentViewController" customModule="Read_and_Tell" customModuleProvider="target" sceneMemberID="viewController">
+                <viewController id="wwr-ya-60d" customClass="ReviewsViewController" customModule="Read_and_Tell" customModuleProvider="target" sceneMemberID="viewController">
                     <view key="view" contentMode="scaleToFill" id="4wy-8U-6nq">
-                        <rect key="frame" x="0.0" y="0.0" width="414" height="838"/>
+                        <rect key="frame" x="0.0" y="0.0" width="414" height="896"/>
                         <autoresizingMask key="autoresizingMask" widthSizable="YES" heightSizable="YES"/>
                         <subviews>
                             <tableView clipsSubviews="YES" contentMode="scaleToFill" fixedFrame="YES" alwaysBounceVertical="YES" dataMode="prototypes" style="plain" separatorStyle="default" rowHeight="-1" estimatedRowHeight="-1" sectionHeaderHeight="-1" estimatedSectionHeaderHeight="-1" sectionFooterHeight="-1" estimatedSectionFooterHeight="-1" translatesAutoresizingMaskIntoConstraints="NO" id="lb0-Qe-XUW">
@@ -650,7 +646,7 @@
                                     <tableViewCell clipsSubviews="YES" contentMode="scaleToFill" preservesSuperviewLayoutMargins="YES" selectionStyle="default" indentationWidth="10" reuseIdentifier="AddReviewCell" id="Uv0-cg-m7M">
                                         <rect key="frame" x="0.0" y="202" width="414" height="43.5"/>
                                         <autoresizingMask key="autoresizingMask"/>
-                                        <tableViewCellContentView key="contentView" autoresizesSubviews="NO" opaque="NO" clipsSubviews="YES" multipleTouchEnabled="YES" contentMode="center" preservesSuperviewLayoutMargins="YES" insetsLayoutMarginsFromSafeArea="NO" tableViewCell="Uv0-cg-m7M" id="ZEu-4I-vmy">
+                                        <tableViewCellContentView key="contentView" opaque="NO" clipsSubviews="YES" multipleTouchEnabled="YES" contentMode="center" preservesSuperviewLayoutMargins="YES" insetsLayoutMarginsFromSafeArea="NO" tableViewCell="Uv0-cg-m7M" id="ZEu-4I-vmy">
                                             <rect key="frame" x="0.0" y="0.0" width="414" height="43.5"/>
                                             <autoresizingMask key="autoresizingMask"/>
                                             <subviews>
@@ -672,12 +668,12 @@
                     </view>
                     <navigationItem key="navigationItem" id="3mo-xF-jHJ"/>
                     <connections>
-                        <outlet property="reviewTableView" destination="lb0-Qe-XUW" id="aTX-JD-LVC"/>
+                        <outlet property="reviewTableView" destination="lb0-Qe-XUW" id="pg1-j2-ocS"/>
                     </connections>
                 </viewController>
                 <placeholder placeholderIdentifier="IBFirstResponder" id="JTg-wZ-JoB" userLabel="First Responder" customClass="UIResponder" sceneMemberID="firstResponder"/>
             </objects>
-            <point key="canvasLocation" x="-178" y="1859"/>
+            <point key="canvasLocation" x="-246.37681159420291" y="1885.7142857142856"/>
         </scene>
         <!--Login View Controller-->
         <scene sceneID="XgF-3o-C7R">
