+++ Read and Tell/BookGridCell.swift
@@ -9,6 +9,7 @@ import UIKit
 
 class BookGridCell: UICollectionViewCell {
     
+    
     @IBOutlet weak var bookTitleLabel: UILabel!
     @IBOutlet weak var bookCoverImage: UIImageView!
