+++ Read and Tell/FeaturedBookDetailViewController.swift
@@ -10,11 +10,18 @@ import UIKit
 class FeaturedBookDetailViewController: UIViewController {
 
     @IBOutlet weak var featuredBookImageView: UIImageView!
+    
+    
     @IBOutlet weak var featuredBookTitleLabel: UILabel!
+    
+    
     @IBOutlet weak var featuredBookAuthorLabel: UILabel!
+    
+    
     @IBOutlet weak var featuredBookDescriptionLabel: UILabel!
     
-    var book: NSDictionary!
+    var book: [String:Any]!
+    
     
     override func viewDidLoad() {
         super.viewDidLoad()
@@ -22,6 +29,7 @@ class FeaturedBookDetailViewController: UIViewController {
         // Do any additional setup after loading the view.
     }
     
+
     /*
     // MARK: - Navigation
