+++ Read and Tell/SearchScreenViewController.swift
@@ -27,7 +27,7 @@ class SearchScreenViewController: UIViewController, UITableViewDataSource, UITab
         self.searchTableView.estimatedRowHeight = 210
         
         searchBar.inputTextView.placeholder = "Search any book by title, author or genre!"
-        searchBar.sendButton.image = UIImage(systemName: "magnifyingglass")
+        searchBar.sendButton.title = "Search"
         searchBar.delegate = self
         
         searchTableView.keyboardDismissMode = .interactive
